move a few more methods
